// pokeApp/src/interfaces/interfacePokemon.tsx
// Generated by https://quicktype.io

export interface PokemonsPaginateResponse {
    count: number;
    next: string;
    previous: null;
    results: Result[];
}

export interface Result {
    name: string;
    url: string;
}



export interface SinglePokemons {
    id: string,
    name: string;
    picture: string;
    color?: string,
}